Updated Identifier

+++ text2text/identifier.py
@@ -1,15 +1,15 @@
+import os
 import requests
 import pickle
 import text2text as t2t
 
 
-class Identifier(t2t.Vectorizer):
+class Identifier():
 
-    def _get_model():
+    def _get_model(self):
         """Download and loads model from google drive"""
 
-        file_id = "1pB1GkWiafx23jmND6psg1cUyBqRq0rVY"
-        model_file = "identifier.pkl"
+        file_id = "10-YNx9yDVJTUijVz_2aopcIF8rdXR78R"
         URL = "https://docs.google.com/uc?export=download"
         session = requests.Session()
         response = session.get(URL, params = { 'id' : file_id }, stream = True)
@@ -19,20 +19,32 @@ class Identifier(t2t.Vectorizer):
                 response = session.get(URL, params = params, stream = True)
 
         CHUNK_SIZE = 32768
-        with open(model_file, "wb") as f:
-            for chunk in response.iter_content(CHUNK_SIZE):
-                if chunk: # filter out keep-alive new chunks
-                    f.write(chunk)
+
+        model_file = "identifier.pkl"
+        if not os.path.exists(model_file):
+            with open(model_file, "wb") as f:
+                for chunk in response.iter_content(CHUNK_SIZE):
+                    if chunk: # filter out keep-alive new chunks
+                        f.write(chunk)
+            
+            model = pickle.load(open(model_file, 'rb'))
         
-        model = pickle.load(open("identifier.pkl", 'rb'))
+        else:
+            model = pickle.load(open(model_file, 'rb'))
+    
         return model
+
 
     def transform(self, input_lines, src_lang='en', **kwargs):
         input_lines = t2t.Transformer.transform(self, input_lines, src_lang=src_lang, **kwargs)
-        identifier = self._get_model()
+        identifier = self.model 
         embeddings = t2t.Vectorizer.transform(self, input_lines, src_lang='en', **kwargs)
         predictions = identifier.predict(embeddings)
         for prediction in predictions:
             print('Language is {}'.format(prediction))
 
         return None
+
+    def __init__(self, **kwargs):
+        self.model = self._get_model()
+
